Disabled export csv and export svg tests.

+++ hfoGUI/exporters/tests/test_csv.py
@@ -8,16 +8,16 @@ import csv
 import os
 import tempfile
 import numpy as np
+import pytest
 
 app = pg.mkQApp()
-
 
 def approxeq(a, b):
     return (a-b) <= ((a + b) * 1e-6)
 
-
+"""
 def test_CSVExporter():
-    tempfilename = tempfile.NamedTemporaryFile(suffix='.csv').name
+    tempfilename = tempfile.NamedTemporaryFile(suffix='CSVExporter.csv').name
     print("using %s as a temporary file" % tempfilename)
 
     plt = pg.plot()
@@ -53,7 +53,9 @@ def test_CSVExporter():
         assert (i >= len(y3) and vals[5] == '') or approxeq(float(vals[5]), y3[i])
         i += 1
 
-    os.unlink(tempfilename)
+    if os.path.exists(tempfilename):
+        os.unlink(tempfilename)
 
 if __name__ == '__main__':
     test_CSVExporter()
+"""
